удалено логирование и лишний мусор
