--- packages/server/config/config.js
@@ -1,2 +1,0 @@
-require('ts-node/register')
-module.exports = require('./config.ts').default
