--- packages/server/models/comment.ts
@@ -1,40 +1,0 @@
-import { DataTypes, Sequelize } from 'sequelize'
-
-const Comment = (sequelize: Sequelize) => {
-  return sequelize.define(
-    'Comment',
-    {
-      id: {
-        type: DataTypes.INTEGER,
-        autoIncrement: true,
-        primaryKey: true,
-      },
-      text: {
-        type: DataTypes.TEXT,
-        allowNull: false,
-      },
-      topicId: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-      },
-      userId: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-      },
-      fileUrl: {
-        type: DataTypes.STRING,
-        allowNull: true,
-      },
-      createdAt: {
-        type: DataTypes.DATE,
-        defaultValue: DataTypes.NOW,
-      },
-    },
-    {
-      tableName: 'comments',
-      timestamps: false,
-    }
-  )
-}
-
-export { Comment }
