--- packages/server/models/reply.ts
@@ -1,40 +1,0 @@
-import { DataTypes, Sequelize } from 'sequelize'
-
-const Reply = (sequelize: Sequelize) => {
-  return sequelize.define(
-    'Reply',
-    {
-      id: {
-        type: DataTypes.INTEGER,
-        autoIncrement: true,
-        primaryKey: true,
-      },
-      text: {
-        type: DataTypes.TEXT,
-        allowNull: false,
-      },
-      commentId: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-      },
-      userId: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-      },
-      fileUrl: {
-        type: DataTypes.STRING,
-        allowNull: true,
-      },
-      createdAt: {
-        type: DataTypes.DATE,
-        defaultValue: DataTypes.NOW,
-      },
-    },
-    {
-      tableName: 'replies',
-      timestamps: false,
-    }
-  )
-}
-
-export { Reply }
