--- packages/server/models/topic.ts
@@ -1,36 +1,0 @@
-import { DataTypes, Sequelize } from 'sequelize'
-
-const Topic = (sequelize: Sequelize) => {
-  return sequelize.define(
-    'Topic',
-    {
-      id: {
-        type: DataTypes.INTEGER,
-        autoIncrement: true,
-        primaryKey: true,
-      },
-      title: {
-        type: DataTypes.STRING,
-        allowNull: false,
-      },
-      userId: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-      },
-      fileUrl: {
-        type: DataTypes.STRING,
-        allowNull: true,
-      },
-      createdAt: {
-        type: DataTypes.DATE,
-        defaultValue: DataTypes.NOW,
-      },
-    },
-    {
-      tableName: 'topics',
-      timestamps: false,
-    }
-  )
-}
-
-export { Topic }
